requirements.txt
Django==1.8.8
django-carton==1.2.0
django-countries==3.4.1
django-crispy-forms==1.5.2
django-debug-toolbar==1.4
django-paypal==0.3.2
djangorestframework==3.3.3
djrill==2.0
gunicorn==19.4.5
Pillow==3.2.0
psycopg2==2.6.1
requests==2.9.1
six==1.10.0
South==1.0.2
sqlparse==0.1.19
